LineChart + Portfolio
Portfiolo erweitert
LineChart etwas generischer. (TemperatureChart -> LineChart)
Variable für die Beschreibung der Linie
